✨ feat: WKPC-62 Refacotred good stuff list into multi-step form
    NOTION: [WKPC-62](https://www.notion.so/work-pace/WKPC-62)
    BREAKING-CHANGE: no
