+++ src/modules/TheGoodStuffList/entries/GeneratedReport/GeneratedReport.module.scss
