+++ src/modules/TheGoodStuffList/entries/MultiStepForm/MultiStepForm.module.scss
@@ -1,0 +1,150 @@
+@use '@workpace/design-system/dist/design-tokens' as *;
+
+.container {
+  padding: $size-200;
+}
+
+.fieldSet {
+  display: flex;
+  flex-flow: row nowrap;
+  width: 100%;
+  justify-content: space-between;
+}
+.fieldSetTitle {
+  word-break: none;
+}
+
+.fields {
+  display: flex;
+  flex-flow: column nowrap;
+  gap: $size-200;
+}
+
+.nav {
+  display: flex;
+  flex-flow: row nowrap;
+  width: 100%;
+  gap: $size-500;
+  padding: $size-300;
+  justify-content: space-between;
+}
+
+.content {
+  margin-bottom: $size-300;
+}
+
+$dim: 40px;
+.circle {
+  width: $dim;
+  height: $dim;
+  background-color: black;
+  border-radius: 50%;
+  color: white;
+  align-items: center;
+  display: flex;
+  justify-content: center;
+}
+
+.active {
+  background-color: $color-active-800;
+}
+
+.actions {
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: space-between;
+}
+
+.step {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: $size-50;
+}
+
+// @use '@workpace/design-system/dist/design-tokens' as *;
+
+.promptWrapper {
+  display: flex;
+  flex-flow: row wrap;
+  gap: $size-150;
+  align-items: flex-end;
+  justify-content: flex-end;
+  outline: solid black 1px;
+  border-radius: $size-50;
+  width: 100%;
+  height: auto;
+  margin: 10% 0;
+}
+
+.textarea {
+  background-color: white;
+  color: black;
+  border: solid gray 1px;
+  padding: 8px;
+  border-radius: 2px;
+
+  flex-grow: 1;
+  min-height: 36px;
+  border: none;
+  resize: none;
+
+  height: auto;
+  overflow-y: hidden; // Prevents scrolling
+  width: 100%;
+
+  &:focus {
+    outline: none;
+  }
+}
+
+.button {
+  height: 24px;
+  border-radius: $size-50;
+  &:hover {
+    background-color: black;
+    color: white;
+  }
+}
+
+.arrow {
+  display: flex;
+  align-items: center;
+  flex-flow: row nowrap;
+  justify-content: space-around;
+  width: 100%;
+  padding: $size-200;
+}
+
+.actions {
+  display: flex;
+  flex-flow: row nowrap;
+  align-items: center;
+  gap: $size-125;
+}
+
+@media #{$mq-sm-only} {
+  .actions {
+    display: flex;
+    flex-flow: row wrap;
+    align-items: center;
+    justify-content: flex-end;
+    gap: $size-125;
+  }
+}
+
+.submitButton {
+  padding-right: $size-100;
+  fill: $color-active-800;
+  &:hover {
+    fill: green;
+  }
+}
+
+.loading {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin: 50px;
+}
