+++ src/modules/TheGoodStuffList/entries/MultiStepForm/index.ts
@@ -1,0 +1,1 @@
+export { MultiStepForm } from './MultiStepForm'
