+++ src/modules/TheGoodStuffList/entries/index.ts
@@ -1,2 +1,3 @@
 export * from './GeneratedReport'
 export * from './AllPages'
+export * from './MultiStepForm'
