+++ src/modules/TheGoodStuffList/widgets/GenerateReportUserPromptInput/GenerateReportUserPromptInput.module.scss
