+++ src/pages/prototypes/good-stuff-list.tsx
@@ -6,7 +6,7 @@ import { getNotionDatabasesController } from '@/api/controllers'
 import { NotionDatabase } from '@/interfaces/notion'
 import { DocumentTitle } from '@/layout/DocumentTitle'
 import GoodStuffList from '@/layout/pages/GoodStuffListPage/GoodStuffListPage'
-import { NotionDatabaseContextProvider } from '@/modules/AccomplishmentReport/contexts'
+import { NotionDatabaseContextProvider } from '@/modules/TheGoodStuffList/contexts'
 import { withNotionClient } from '@/server/utils/withNotionClient'
 import { withPageRequestWrapper } from '@/server/utils/withPageRequestWrapper'
 import HeroLayout from '@/layout/HeroLayout'
